Refactor: Main, Feed 컴포넌트 세부 컴포넌트로 분리
