+++ src/pages/donghokim/Main/Components/Aside/Aside.scss
@@ -1,0 +1,149 @@
+@import "../../../../../styles/common.scss";
+@import "../../../styles/base.scss";
+@import "../ThumbnailImage/ThumbnailImage.scss";
+
+@media screen and (max-width: 999px) {
+  .main_right {
+    display: none;
+  }
+}
+
+.main_right {
+  max-width: 293px;
+  width: 100%;
+  margin-top: 10px;
+  margin-left: 10px;
+
+  .user_container {
+    display: flex;
+    max-height: 141px;
+    margin-top: 18px;
+    margin-bottom: 20px;
+
+    .nickname {
+      @include textButton();
+    }
+
+    .thumbnail_image_container {
+      @include ThumbnailImage(56, false);
+    }
+
+    .user_info_container {
+      display: flex;
+      flex-direction: column;
+      justify-content: center;
+      align-items: flex-start;
+      margin-left: 10px;
+    }
+
+    .main_user {
+      margin-left: 16px;
+    }
+
+    .nickname {
+      margin-bottom: 2px;
+    }
+
+    .username {
+      margin-top: 2px;
+      font-size: 12px;
+      color: rgb(142, 142, 142);
+      line-height: 14px;
+      overflow: hidden;
+      text-overflow: ellipsis;
+      white-space: nowrap;
+      letter-spacing: 0.1px;
+    }
+  }
+
+  .recommendation_containter {
+    background: rgb(250, 250, 250);
+
+    .recommendation_text {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      font-weight: 600;
+      font-size: 14px;
+      color: rgb(142, 142, 142);
+
+      .show_all_button {
+        @include textButton();
+        font-size: 12px;
+      }
+    }
+
+    .recommended_users {
+      display: flex;
+      flex-direction: column;
+      margin-left: 4px;
+      margin-top: 8px;
+      margin-bottom: 15px;
+
+      .recommended_user {
+        display: flex;
+        justify-content: space-between;
+        height: 48px;
+
+        .user_container {
+          display: flex;
+          align-items: center;
+
+          .thumbnail_image_container {
+            @include ThumbnailImage(32, false);
+          }
+
+          .user_info_container {
+            .description {
+              font-size: 12px;
+              line-height: 14px;
+              color: rgb(142, 142, 142);
+            }
+          }
+        }
+
+        .follow_button {
+          @include textButton();
+          font-size: 12px;
+          color: rgb(0, 149, 246);
+        }
+      }
+    }
+  }
+
+  .info_container {
+    .info_nav {
+      max-width: 100%;
+      margin-bottom: 16px;
+
+      ul {
+        margin-bottom: 3px;
+
+        li {
+          display: inline-block;
+          font-size: 11px;
+          font-weight: 400;
+          line-height: 13px;
+          color: rgb(199, 199, 199);
+
+          a {
+            color: inherit;
+            text-decoration: none;
+          }
+
+          &:not(:last-child)::after {
+            content: " ∙ ";
+          }
+        }
+      }
+    }
+
+    .copyright {
+      font-size: 11px;
+      font-weight: 400;
+      line-height: 13px;
+      color: rgb(199, 199, 199);
+      text-transform: uppercase;
+    }
+  }
+}
