+++ src/pages/donghokim/Main/Components/Buttons/index.js
@@ -1,0 +1,6 @@
+export * from "./CommentButton";
+export * from "./HeartButton";
+export * from "./LikeButton";
+export * from "./OptionButton";
+export * from "./ShareButton";
+export * from "./StoreButton";
