+++ src/pages/donghokim/Main/Components/CommentContainer/CommentsContainer.js
@@ -1,5 +1,6 @@
 import React from "react";
 import Comment from "./Comment";
+import "./CommentsContainer.scss";
 
 class CommentsContainer extends React.Component {
   render() {
