+++ src/pages/donghokim/Main/Components/CommentContainer/CommentsContainer.scss
@@ -1,0 +1,38 @@
+@import "../../../../../styles/common.scss";
+@import "../../../styles/base.scss";
+
+.comment_container {
+  display: flex;
+  justify-content: space-between;
+  padding-left: 15px;
+  padding-right: 15px;
+  margin-bottom: 2px;
+
+  .comment {
+    height: 22px;
+    width: 100%;
+
+    button {
+      @include textButton();
+    }
+
+    .text {
+      margin-left: 8px;
+      font-size: 14px;
+    }
+  }
+
+  button {
+    border: 0;
+    outline: none;
+    background: transparent;
+  }
+
+  .option_button {
+    opacity: 0;
+    position: absolute;
+    right: 40px;
+    cursor: pointer;
+    transition: opacity;
+  }
+}
