+++ src/pages/donghokim/Main/Components/Navbar/Navbar.js
@@ -1,5 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
+import "./Navbar.scss";
 
 class Navbar extends React.Component {
   render() {
