+++ src/pages/donghokim/Main/Components/Navbar/Navbar.scss
@@ -1,0 +1,111 @@
+@import "../../../../../styles/common.scss";
+@import "../../../styles/base.scss";
+
+.navbar {
+  position: fixed;
+  top: 0;
+  left: 0;
+  display: flex;
+  justify-content: center;
+  height: 54px;
+  width: 100%;
+  background: $white;
+  border-bottom: 1px solid $border-grey;
+  z-index: 2;
+
+  .navbar_container {
+    @include centerSpaceBetween;
+    max-width: 975px;
+    width: 100%;
+
+    .logo_container {
+      position: relative;
+      top: 5px;
+      left: 18px;
+    }
+
+    .searchbar_container {
+      position: relative;
+      left: 48px;
+      height: 28px;
+      min-width: 125px;
+      width: 215px;
+      cursor: text;
+
+      input {
+        height: 70%;
+        width: 85%;
+        font-size: 16px;
+        background: $background-grey;
+        border: 1px solid $border-grey;
+        border-radius: 3px;
+        color: $text-black;
+        outline: none;
+        padding: 3px 10px 3px 26px;
+        z-index: 2;
+      }
+
+      .placeholder {
+        position: absolute;
+        top: 3px;
+        left: -8px;
+        display: flex;
+        justify-content: center;
+        height: 100%;
+        width: 100%;
+        padding: 4px;
+        opacity: 0.8;
+
+        .searchbar_icon {
+          height: 10px;
+          width: 10px;
+          display: inline-block;
+          margin-top: 3px;
+          margin-right: 6px;
+          background: url("https://instagram.com/static/bundles/es6/sprite_core_32f0a4f27407.png/32f0a4f27407.png");
+          background-repeat: no-repeat;
+          background-position: -399px -321px;
+        }
+
+        .searchbar_text {
+          font-size: 14px;
+          font-weight: 200;
+          color: $text-grey;
+          text-align: center;
+        }
+      }
+    }
+
+    .navbutton_container {
+      position: relative;
+      left: -4px;
+      top: 1px;
+      @include centerSpaceBetween;
+      flex-wrap: wrap;
+      width: 200px;
+      margin-right: 12px;
+
+      .button {
+        height: 22px;
+
+        a {
+          color: rgb(0, 55, 107);
+          text-decoration: none;
+        }
+
+        .user_setting {
+          display: block;
+          height: 22px;
+          width: 22px;
+          cursor: pointer;
+
+          img {
+            height: 100%;
+            width: 100%;
+            border-radius: 50%;
+          }
+        }
+      }
+    }
+  }
+}
