--- src/pages/donghokim/Main/Storybar.js
@@ -1,139 +1,0 @@
-import React from "react";
-
-class Storybar extends React.Component {
-  render() {
-    return (
-      <div className="storybar_slider_wrap">
-        <div className="storybar_slider_view">
-          <div className="storybar_slider_lists">
-            <div className="user_story">
-              <button>
-                <div className="thumbnail_image_container">
-                  <img
-                    alt="user_thumbnail"
-                    src="https://scontent-gmp1-1.cdninstagram.com/v/t51.2885-19/s150x150/67245341_415311355999011_313920263471235072_n.jpg?_nc_ht=scontent-gmp1-1.cdninstagram.com&_nc_ohc=QzZa1aemIuoAX_FyVwr&oh=951ebbbac430f1d5996b57647ed22f89&oe=5F4ECA80"
-                    height="56px"
-                    width="56px"
-                  />
-                </div>
-                <span className="nickname">heavy_nny</span>
-              </button>
-            </div>
-            <div className="user_story">
-              <button>
-                <div className="thumbnail_image_container">
-                  <img
-                    alt="user_thumbnail"
-                    src="https://scontent-gmp1-1.cdninstagram.com/v/t51.2885-19/s150x150/67245341_415311355999011_313920263471235072_n.jpg?_nc_ht=scontent-gmp1-1.cdninstagram.com&_nc_ohc=QzZa1aemIuoAX_FyVwr&oh=951ebbbac430f1d5996b57647ed22f89&oe=5F4ECA80"
-                    height="56px"
-                    width="56px"
-                  />
-                </div>
-                <span className="nickname">heavy_nny</span>
-              </button>
-            </div>
-            <div className="user_story">
-              <button>
-                <div className="thumbnail_image_container">
-                  <img
-                    alt="user_thumbnail"
-                    src="https://scontent-gmp1-1.cdninstagram.com/v/t51.2885-19/s150x150/67245341_415311355999011_313920263471235072_n.jpg?_nc_ht=scontent-gmp1-1.cdninstagram.com&_nc_ohc=QzZa1aemIuoAX_FyVwr&oh=951ebbbac430f1d5996b57647ed22f89&oe=5F4ECA80"
-                    height="56px"
-                    width="56px"
-                  />
-                </div>
-                <span className="nickname">heavy_nny</span>
-              </button>
-            </div>
-            <div className="user_story">
-              <button>
-                <div className="thumbnail_image_container">
-                  <img
-                    alt="user_thumbnail"
-                    src="https://scontent-gmp1-1.cdninstagram.com/v/t51.2885-19/s150x150/67245341_415311355999011_313920263471235072_n.jpg?_nc_ht=scontent-gmp1-1.cdninstagram.com&_nc_ohc=QzZa1aemIuoAX_FyVwr&oh=951ebbbac430f1d5996b57647ed22f89&oe=5F4ECA80"
-                    height="56px"
-                    width="56px"
-                  />
-                </div>
-                <span className="nickname">heavy_nny</span>
-              </button>
-            </div>
-            <div className="user_story">
-              <button>
-                <div className="thumbnail_image_container">
-                  <img
-                    alt="user_thumbnail"
-                    src="https://scontent-gmp1-1.cdninstagram.com/v/t51.2885-19/s150x150/67245341_415311355999011_313920263471235072_n.jpg?_nc_ht=scontent-gmp1-1.cdninstagram.com&_nc_ohc=QzZa1aemIuoAX_FyVwr&oh=951ebbbac430f1d5996b57647ed22f89&oe=5F4ECA80"
-                    height="56px"
-                    width="56px"
-                  />
-                </div>
-                <span className="nickname">heavy_nny</span>
-              </button>
-            </div>
-            <div className="user_story">
-              <button>
-                <div className="thumbnail_image_container">
-                  <img
-                    alt="user_thumbnail"
-                    src="https://scontent-gmp1-1.cdninstagram.com/v/t51.2885-19/s150x150/67245341_415311355999011_313920263471235072_n.jpg?_nc_ht=scontent-gmp1-1.cdninstagram.com&_nc_ohc=QzZa1aemIuoAX_FyVwr&oh=951ebbbac430f1d5996b57647ed22f89&oe=5F4ECA80"
-                    height="56px"
-                    width="56px"
-                  />
-                </div>
-                <span className="nickname">heavy_nny</span>
-              </button>
-            </div>
-            <div className="user_story">
-              <button>
-                <div className="thumbnail_image_container">
-                  <img
-                    alt="user_thumbnail"
-                    src="https://scontent-gmp1-1.cdninstagram.com/v/t51.2885-19/s150x150/67245341_415311355999011_313920263471235072_n.jpg?_nc_ht=scontent-gmp1-1.cdninstagram.com&_nc_ohc=QzZa1aemIuoAX_FyVwr&oh=951ebbbac430f1d5996b57647ed22f89&oe=5F4ECA80"
-                    height="56px"
-                    width="56px"
-                  />
-                </div>
-                <span className="nickname">heavy_nny</span>
-              </button>
-            </div>
-            <div className="user_story">
-              <button>
-                <div className="thumbnail_image_container">
-                  <img
-                    alt="user_thumbnail"
-                    src="https://scontent-gmp1-1.cdninstagram.com/v/t51.2885-19/s150x150/67245341_415311355999011_313920263471235072_n.jpg?_nc_ht=scontent-gmp1-1.cdninstagram.com&_nc_ohc=QzZa1aemIuoAX_FyVwr&oh=951ebbbac430f1d5996b57647ed22f89&oe=5F4ECA80"
-                    height="56px"
-                    width="56px"
-                  />
-                </div>
-                <span className="nickname">heavy_nny</span>
-              </button>
-            </div>
-            <div className="user_story">
-              <button>
-                <div className="thumbnail_image_container">
-                  <img
-                    alt="user_thumbnail"
-                    src="https://scontent-gmp1-1.cdninstagram.com/v/t51.2885-19/s150x150/67245341_415311355999011_313920263471235072_n.jpg?_nc_ht=scontent-gmp1-1.cdninstagram.com&_nc_ohc=QzZa1aemIuoAX_FyVwr&oh=951ebbbac430f1d5996b57647ed22f89&oe=5F4ECA80"
-                    height="56px"
-                    width="56px"
-                  />
-                </div>
-                <span className="nickname">heavy_nny</span>
-              </button>
-            </div>
-          </div>
-          <button className="left_button">
-            <div className="button_image"></div>
-          </button>
-          <button className="right_button">
-            <div className="button_image"></div>
-          </button>
-        </div>
-      </div>
-    );
-  }
-}
-
-export default Storybar;
